% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{competencies_flatten}
\alias{competencies_flatten}
\title{Pull from Competencies Flatten API endpoint.}
\usage{
competencies_flatten(text, token, bias = NULL, exclude = NULL, shush = FALSE)
}
\arguments{
\item{text}{String. Sentence or text entry used to pull competencies.}

\item{token}{String. Authorization token.}

\item{bias}{Numeric. Level of bias.}

\item{exclude}{Character vector. Vector of result elements to omit.}

\item{shush}{Logical. Print query status.}
}
\description{
Pull from Competencies Flatten API endpoint.
}
\examples{
\dontrun{
txt <- "i am a welder and used welding equipment and acetylene torches"
c <- competencies_flatten(text = txt, bias = 0.4)
}
}
