% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{competencies_military}
\alias{competencies_military}
\title{Pull from Competencies Military API endpoint.}
\usage{
competencies_military(
  moc,
  service_branch = NULL,
  mpc = NULL,
  exclude = NULL,
  start_date = NULL,
  end_date = NULL,
  moc_status = NULL,
  token,
  shush = FALSE
)
}
\arguments{
\item{moc}{String or numeric. MOC to return competencies for.}

\item{service_branch}{String. Branch of Armed Forces.}

\item{mpc}{String.}

\item{exclude}{Character vector. Vector of return elements to exclude.}

\item{start_date}{String. Start date of military service.}

\item{end_date}{String. End date of military service.}

\item{moc_status}{String. Status of MOC.}

\item{token}{String. Authorization token.}

\item{shush}{Logical. Print query status.}
}
\description{
Pull from Competencies Military API endpoint.
}
\examples{
\dontrun{
c <- competencies_military(moc = "2A554E")
}
}
