% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{freshness}
\alias{freshness}
\title{Pull the freshness tag of the data}
\usage{
freshness(token, shush = FALSE)
}
\arguments{
\item{token}{String. Authorization token.}

\item{shush}{Logical. Whether to print request status.}
}
\description{
Pull the freshness tag of the data
}
