% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization_functions.R
\name{get_access_token}
\alias{get_access_token}
\title{Get SkillsEngine Access Token}
\usage{
get_access_token(client_id, client_secret, shush = FALSE)
}
\arguments{
\item{client_id}{String. Provided by \code{get_id_secret}}

\item{client_secret}{String. Provided by \code{get_id_secret}}

\item{shush}{Logical. Print query status or not}
}
\description{
Get SkillsEngine Access Token
}
