% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization_functions.R
\name{get_id_secret}
\alias{get_id_secret}
\title{Get SkillsEngine Client ID and Client Secret}
\usage{
get_id_secret()
}
\value{
List of length 2; Contains Client ID and Secret keys
}
\description{
Get SkillsEngine Client ID and Client Secret
}
