% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{occupations}
\alias{occupations}
\title{Pull the full list of occupations}
\usage{
occupations(token, shush = FALSE)
}
\arguments{
\item{token}{String. Authorization token.}

\item{shush}{Logical. Whether to print request status.}
}
\description{
Pull the full list of occupations
}
