% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{occupations_soc_codes}
\alias{occupations_soc_codes}
\title{Provide economic data related to user-provided SOC codes (by state).}
\usage{
occupations_soc_codes(token, soc_codes, state_code = NULL, shush = FALSE)
}
\arguments{
\item{token}{String. Authorization token.}

\item{soc_codes}{Character vector. SOC codes to pull data for.}

\item{state_code}{String. State code to get more specific results. (Optional)}

\item{shush}{Logical. Whether to print request status.}
}
\description{
Provide economic data related to user-provided SOC codes (by state).
}
\examples{
\dontrun{
o <- occupations_soc_codes(soc_codes = c("13-2011","15-2031","15-2011"),
                           state_code = "TX")
}
}
