% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization_functions.R
\name{set_id_secret}
\alias{set_id_secret}
\title{Set SkillsEngine Client ID and Client Secret}
\usage{
set_id_secret()
}
\value{
No value; called for side effects
}
\description{
Set SkillsEngine Client ID and Client Secret
}
