% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{skills}
\alias{skills}
\title{Pull from Skills API endpoint.}
\usage{
skills(
  text,
  depth = NULL,
  cutoff = NULL,
  socs = NULL,
  similarity_scoring = NULL,
  token,
  shush = FALSE
)
}
\arguments{
\item{text}{String. Sentence or text entry used to pull competencies.}

\item{depth}{Numeric. Algorithm depth.}

\item{cutoff}{Numeric. Cutoff value.}

\item{socs}{Character vector. Vector of SOCs to fine-tune search.}

\item{similarity_scoring}{String. Whether or not to use similarity scoring.}

\item{token}{String. Authorization token.}

\item{shush}{Logical. Print query status.}
}
\description{
Pull from Skills API endpoint.
}
\examples{
\dontrun{
txt <- "Provide substance abuse education and counseling for at-risk individuals."
s <- skills(text = txt,
            depth = 20,
            cutoff = 5,
            similarity_scoring = "true")
}
}
