% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{skills_multi_match}
\alias{skills_multi_match}
\title{Pull from Skills Multi-Match API endpoint.}
\usage{
skills_multi_match(
  sentences,
  depth = NULL,
  cutoff = NULL,
  socs = NULL,
  similarity_scoring = NULL,
  token,
  shush = FALSE
)
}
\arguments{
\item{sentences}{Character vector. Sentences used to pull competencies.}

\item{depth}{Numeric. Algorithm depth.}

\item{cutoff}{Numeric. Cutoff value.}

\item{socs}{Character vector. Vector of SOCs to fine-tune search.}

\item{similarity_scoring}{String. Whether or not to use similarity scoring.}

\item{token}{String. Authorization token.}

\item{shush}{Logical. Print query status.}
}
\description{
Pull from Skills Multi-Match API endpoint.
}
\examples{
\dontrun{
txts <- c("Weld things together", "Use an acetylene torch", "Hammer nails into wood")
s <- skills_multi_match(sentences = txts,
                        depth = 20,
                        cutoff = 5,
                        socs = c("17-2141.00",
                                 "49-9041.00",
                                 "49-1011.00",
                                 "49-9071.00"))
}
}
