% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{test_access}
\alias{test_access}
\title{Test connection to SkillsEngine API}
\usage{
test_access(token)
}
\arguments{
\item{token}{String. Authorization token.}
}
\description{
Test connection to SkillsEngine API
}
