% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_functions.R
\name{topics_related}
\alias{topics_related}
\title{Pull from Related Topics API endpoint.}
\usage{
topics_related(texts, socs = NULL, token, shush = FALSE)
}
\arguments{
\item{texts}{Character Vector. Sentences used to pull related topics.}

\item{socs}{Character vector. Soc codes to fine-tune related topics.}

\item{token}{String. Authorization token.}

\item{shush}{Logical. Print query status.}
}
\description{
Pull from Related Topics API endpoint.
}
\examples{
\dontrun{
t <- topics_related(texts = c("ms word", "access", "dexterity"),
                    socs = "41-3041.00")
}
}
