% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{freshness}
\alias{freshness}
\title{Pull the freshness tag of the data}
\usage{
freshness(token, handle_status = "warn")
}
\arguments{
\item{token}{Authorization token obtained from \code{get_access_token}}

\item{handle_status}{How to handle bad HTTP status. Set as either \code{warn} or \code{error}}
}
\value{
A string; the data freshness tag
}
\description{
Pull the freshness tag of the data
}
