% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization_functions.R
\name{get_access_token}
\alias{get_access_token}
\title{Get SkillsEngine Access Token}
\usage{
get_access_token(client_id, client_secret, handle_status = "warn")
}
\arguments{
\item{client_id}{Client ID provided by \code{get_id_secret}, or manually entered}

\item{client_secret}{Client Secret provided by \code{get_id_secret}, or manually entered}

\item{handle_status}{How to handle bad HTTP status. Set as either \code{warn} or \code{error}}
}
\value{
A string; the Access Token
}
\description{
Get SkillsEngine Access Token
}
