% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{occupations}
\alias{occupations}
\title{Pull the full list of occupations}
\usage{
occupations(token, handle_status = "warn", response_raw = FALSE)
}
\arguments{
\item{token}{Authorization token obtained from \code{get_access_token}}

\item{handle_status}{How to handle bad HTTP status. Set as either \code{warn} or \code{error}}

\item{response_raw}{Logical value whether to return the API response as
raw, unparsed text. Defaults to \code{FALSE}}
}
\value{
A data.frame; occupations relevant to query
}
\description{
Pull the full list of occupations
}
