% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{test_access}
\alias{test_access}
\title{Test connection to SkillsEngine API}
\usage{
test_access(token, handle_status = "warn")
}
\arguments{
\item{token}{Authorization token obtained from \code{get_access_token}}

\item{handle_status}{How to handle bad HTTP status. Set as either \code{warn} or \code{error}}
}
\value{
No return value; Prints status to console
}
\description{
Test connection to SkillsEngine API
}
